% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/call_snp.R
\name{call_snp}
\alias{call_snp}
\title{call SNPs}
\usage{
call_snp(x)
}
\arguments{
\item{x}{pileup dataframe}
}
\description{
R function to call SNPs from a pileup dataframe
}
\examples{
call_snp(x)
}
\keyword{Pileup}

