% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pca.R
\name{draw_pca}
\alias{draw_pca}
\title{Draw PCA}
\usage{
draw_pca(x, point_names, pc_x, pc_y)
}
\arguments{
\item{x}{dataframe}

\item{point_names}{name to use as name of points [string]}

\item{pc_x}{x-axis pc [string]}

\item{pc_y}{y-axis pc [string]}
}
\description{
Draw generic PCA of data
}

