% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pca_2.R
\name{draw_pca_2}
\alias{draw_pca_2}
\title{Draw PCA 2}
\usage{
draw_pca_2(x, ids = "XXX")
}
\arguments{
\item{x}{dataframe}

\item{id}{sample id ["XXX"]}
}
\description{
R function draw a PCA from an evec dataframe
}
\examples{
draw_pca_2(x, ids="sample1")
}

