% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pca_laser.R
\name{draw_pca_laser}
\alias{draw_pca_laser}
\title{Draw PCA laser}
\usage{
draw_pca_laser(x, y, bam_file_string = "-merge-rmdup", size_of_name = 3)
}
\arguments{
\item{x}{dataframe of reference data}

\item{y}{dataframe of sample data}

\item{bam_file_string}{string to remove from sample names}

\item{size_of_name}{size of name to plot}
}
\description{
R function draw a PCA from an evec dataframe
}

