% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pca_names.R
\name{draw_pca_names}
\alias{draw_pca_names}
\title{Draw PCA names}
\usage{
draw_pca_names(x, ids = "XXX", size_of_name = 3)
}
\arguments{
\item{x}{dataframe of reference data}

\item{ids}{sample id ["XXX"]}

\item{size_of_name}{size of name to plot}
}
\description{
R function draw a PCA from an evec dataframe
}

