% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_evec.R
\name{filter_evec}
\alias{filter_evec}
\title{Filter evec}
\usage{
filter_evec(evec_file)
}
\arguments{
\item{evec_file}{input evec file}
}
\description{
R function to collect evec file into a dataframe
}
\examples{
filter_evec(evec_file)
}
\keyword{smartpca}

