% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filter_ped}
\alias{filter_ped}
\title{Draw PCA}
\usage{
filter_ped(input_file_name, output_file_name)
}
\arguments{
\item{input_file}{input file ped}

\item{ouput_file}{output file ped}
}
\description{
R function to make a ped file homzygous
}
\examples{
filter_ped("input_file.ped", "output_file.ped")
}
\keyword{plink}

