% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_pca_evec.R
\name{int_pca_evec}
\alias{int_pca_evec}
\title{Int PCA evec}
\usage{
int_pca_evec(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Experimental R function draw an interactive PCA from a smartpca/evec dataframe
}

