% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Make map}
\usage{
make_map(x, outfile)
}
\arguments{
\item{x}{pileup data frame}

\item{outfile}{- output file}
}
\description{
R function to make map files from a pileup dataframe
}
\examples{
make_map(pileup_df, "output.test")
}
\keyword{Pileup}

