% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_plink.R
\name{make_plink}
\alias{make_plink}
\title{make_plink}
\usage{
make_plink(pileup_file, sample_name)
}
\arguments{
\item{pileup_file}{input pileup file}

\item{sample_name}{imput sample name}
}
\description{
R function to make PED and map files from a pileup dataframe
}
\examples{
make_plink(evec_df, sample_id="sample1")
}
\keyword{Pileup}

