% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pileup_filter.R
\name{pileup_filter}
\alias{pileup_filter}
\title{Pileup filter}
\usage{
pileup_filter(pileup_file)
}
\arguments{
\item{pileup_file}{pileup}
}
\description{
R function filter a pileup file into a dataframe
}
\examples{
call_snp(x)
}
\keyword{Pileup}

