\name{rPCA-package}
\alias{rPCA-package}
\alias{rPCA}
\docType{package}
\title{
Draw PCAs from smartpca and LASER file
}
\description{
Draw PCAs from smartpca and LASER file
}
\details{
\tabular{ll}{
Package: \tab rPCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-07-21\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Matthew Teasdale

Maintainer: Matthew Teasdale <matthew@teasdale.xyz>
}
\references{
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
}
\examples{
}
