% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pca.R
\name{draw_pca}
\alias{draw_pca}
\title{Draw PCA}
\usage{
draw_pca(x, ids = "XXX")
}
\arguments{
\item{x}{dataframe}

\item{ids}{sample id ["XXX"]}
}
\description{
R function draw a PCA from an evec dataframe
}
\examples{
draw_pca(x, ids="sample1")
}

