% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/draw_pca.R
\name{draw_pca}
\alias{draw_pca}
\title{Draw PCA}
\usage{
draw_pca(evec_df, sample_id = "XXX")
}
\arguments{
\item{evec_df}{input dataframe}

\item{sample_id}{["XXX"]}
}
\description{
R function draw a PCA from an evec dataframe
}
\examples{
draw_pca(evec_df, sample_id="sample1")
}
\keyword{smartpca}

