% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pca_laser.R
\name{draw_pca_laser}
\alias{draw_pca_laser}
\title{Draw PCA laser}
\usage{
draw_pca_laser(x)
}
\arguments{
\item{x}{dataframe of reference data}
}
\description{
R function draw a PCA from an evec dataframe
}

