% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_evec}
\alias{filter_evec}
\title{Filter evec}
\usage{
filter_evec(evec_file)
}
\arguments{
\item{evec_file}{input evec file}
}
\description{
R function to collect evec file into a dataframe
}
\examples{
filter_evec(evec_file)
}
\keyword{smartpca}

