\name{rPCA-package}
\alias{rPCA-package}
\alias{rPCA}
\docType{package}
\title{
R package to help with PCAs
}
\description{
R PCA provides some R functions to help with PCAs and the output from smartpca (Patterson et al 2012)
}
\details{
\tabular{ll}{
Package: \tab rPCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-24\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Matthew Teasdale

Maintainer: Matthew Teasdale <matt@teasdale.xyz>
}
\references{
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
}
\examples{
}
